+++ src/components/AppRepoForm/index.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { Redirect } from "react-router";
 
-import { AppRepository } from "../shared/AppRepository";
+import { AppRepository } from "../../shared/AppRepository";
 
 interface IAppRepoFormProps {
   name: string;
