+++ src/components/ServiceBrokerView/index.tsx
@@ -1,5 +1,6 @@
 import * as React from "react";
-import { IServiceBroker } from "../shared/ServiceCatalog";
+
+import { IServiceBroker } from "../../shared/ServiceCatalog";
 
 interface IServiceBrokerViewProps {
   broker: IServiceBroker;
